+++ lib/active_model/serializer.rb
@@ -21,7 +21,8 @@ module ActiveModel
     include Caching
     include Links
     include Type
-    require 'active_model/serializer/adapter'
+    # Deprecated
+    require 'active_model_serializers/adapter'
 
     # @param resource [ActiveRecord::Base, ActiveModelSerializers::Model]
     # @return [ActiveModel::Serializer]
@@ -40,9 +41,11 @@ module ActiveModel
       end
     end
 
-    # @see ActiveModel::Serializer::Adapter.lookup
+    # @see ActiveModelSerializers::Adapter.lookup
+    # Deprecated
     def self.adapter
-      ActiveModel::Serializer::Adapter.lookup(config.adapter)
+      warn 'Calling adapter method in Serializer, please use the ActiveModelSerializers::configured_adapter'
+      ActiveModelSerializers::Adapter.lookup(config.adapter)
     end
 
     # @api private